django/requirements.txt
aliyun-python-sdk-core==2.13.33
aliyun-python-sdk-ram==3.2.0
amqp==2.6.1
asgiref==3.3.1
autopep8==1.5.4
billiard==3.6.3.0
cachetools==4.2.1
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.1.5
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-extensions==3.1.0
django-gitlab-auth==0.1.1
django-guardian==2.3.0
django-kubernetes-manager==0.4.7
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
flower==0.9.7
google-auth==1.24.0
humanize==3.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==4.6.11
kubernetes==12.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
orjson==3.4.7
packaging==20.9
prometheus-client==0.8.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-gitlab==2.6.0
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.3
uWSGI==2.0.19.1
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.57.0
